requirements.txt
# Core AI/LLM libraries
anthropic>=0.18.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langsmith>=0.1.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Web UI
streamlit>=1.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

# Optional: For advanced features in later weeks
# Uncomment as needed

# Vector stores (Week 10-11)
# chromadb>=0.4.0
# faiss-cpu>=1.7.4

# Async support
# aiohttp>=3.9.0

# Logging and monitoring
# structlog>=24.0.0